fix: cast b_do to float32 in _chunk_bwd_dh_kernel dot (#110)
* Initial plan

* fix: cast b_do to float32 in _chunk_bwd_dh_kernel dot operation

Agent-Logs-Url: https://github.com/primatrix/pallas-kernel/sessions/e2a1f139-8e9a-4c70-b8f3-3e845032dd2b


---------

+++ tops/ops/common/chunk_h.py
@@ -761,7 +761,7 @@ def _chunk_bwd_dh_kernel(
         b_q = (b_q * exp(b_gk)).astype(b_q.dtype)
 
     b_dh = b_dh + jax.lax.dot(
-        b_q.astype(jnp.float32).T, b_do,
+        b_q.astype(jnp.float32).T, b_do.astype(jnp.float32),
         precision=lax.Precision.HIGHEST,
         preferred_element_type=jnp.float32,
     )
